我的现金红包页面解决提现按钮判断

+++ MaiBaiApplication/TianShenCash/src/main/java/com/tianshen/cash/activity/RedPackageActivity.kt
@@ -167,7 +167,8 @@ class RedPackageActivity : BaseActivity() {
         var min_withdrawals = MoneyUtils.changeF2Y(data?.min_withdrawals, 2)
 
         //设置提现按钮是否可以点击
-        tv_get_red_package.isEnabled = withdrawals_money > min_withdrawals
+
+        tv_get_red_package.isEnabled = withdrawals_money.toFloat() > min_withdrawals.toFloat()
 
         tv_withdrawals_money.text = withdrawals_money
         tv_all_income.text = all_income
